feat: insert data in batches

+++ src/meta/mod.rs
@@ -24,7 +24,7 @@ pub struct JavaMetaData {
     pub sha256_url: Option<String>,
     pub sha512: Option<String>,
     pub sha512_url: Option<String>,
-    pub size: Option<u64>,
+    pub size: Option<i32>,
     pub url: String,
     pub vendor: String,
     pub version: String,
@@ -50,6 +50,7 @@ impl JavaMetaData {
         map
     }
 }
+
 #[cfg(test)]
 mod tests {
     use super::*;
